backend/requirements.txt
# Core
llama-index>=0.11.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0

# LLM Providers
llama-index-llms-anthropic>=0.3.0
llama-index-llms-openai>=0.2.0
llama-index-llms-azure-openai>=0.2.0
llama-index-llms-ollama>=0.3.0
anthropic>=0.30.0
openai>=1.0.0

# Embeddings
llama-index-embeddings-huggingface>=0.3.0

# Vector Stores
llama-index-vector-stores-chroma>=0.2.0
llama-index-vector-stores-mongodb>=0.2.0
chromadb>=0.5.0
pymongo>=4.6.0

# Hybrid Search (BM25)
llama-index-retrievers-bm25>=0.3.0
rank-bm25>=0.2.2

# Utils
httpx<0.28
